web/requirements.txt
openai-agents
python-dotenv
jsonschema
redis
a2a-sdk
uvicorn